Fix DUMP command reply
The DUMP command returns binary data or '(nil)'.
When redis-py process the response, it tries to decode it to UTF-8 which
cases a python exception.

The DUMP command should not be decoded for this reason, so before
sending it to redis-py we add the `NEVER_DECODE` option.
Now when we get it back we are getting a byte array that looks similar
to this: `b'"\x00\xc0\x0c\n\x00<*\xf8Q\\\x04\xbc\xd5"'`.
In order to look like redis-cli we trim it so it will look like this:
`"\x00\xc0\x0c\n\x00<*\xf8Q\\\x04\xbc\xd5"`.